content/rendering-with-batch/files/cdk-project/requirements.txt
aws-cdk-lib==2.135.0
constructs==10.3.0
cdk-use-cases-custom-cloud9-ssm==1.3.0
